Switching this project back over to Javascript
And using the code from html5-experiments.
My rationale for this mostly that the code was already written and switching over to ruby wasn't seeming to give me enough benefits to outweigh the trouble of trying to learn ruby.
This was probably naive and the reason i was having too much trouble with ruby is because I was trying to bite off more than I could chew, but oh well.
I'll tag this and the previous commit so if I or you, dear reader, ever want to go back and give the ruby version another shot you can go ahead.